requirements/base.txt
# Base requirements
Django>=4.2,<5.0
graphene-django>=3.0
django-filter>=23.1
django-cors-headers>=3.14
psutil>=5.9
django-environ>=0.11
jwt==1.3.1
PyJWT==2.9.0
pyotp==2.9.0
qrcode==7.4.2
rail-django-graphql @ git+https://github.com/raillogistic/rail-django-graphql.git
django-extensions
psycopg2==2.9.9
psycopg2-binary==2.9.9
openpyxl
django-polymorphic